chore: type check fixes
